Refactor aggregate replacement of switch outputs
This essentially pauses aggregate replacement at the end of each switch branch, to be continued when all branches are ready. This way we know how many outputs will be needed when the switch output type is a slice and different branches might resolve to a root array alloca of a different length.